Multi-threaded command recording

+++ Source/Renderer/SwapChain.h
@@ -75,7 +75,7 @@ public:
 
 
 	/* Getters */ 
-	inline int                           GetNumBackBuffers()                const { return mNumBackBuffers; }
+	inline unsigned short                GetNumBackBuffers()                const { return mNumBackBuffers; }
 	inline int                           GetCurrentBackBufferIndex()        const { return mICurrentBackBuffer; }
 	inline CD3DX12_CPU_DESCRIPTOR_HANDLE GetCurrentBackBufferRTVHandle()    const { return CD3DX12_CPU_DESCRIPTOR_HANDLE(mpDescHeapRTV->GetCPUDescriptorHandleForHeapStart(), GetCurrentBackBufferIndex(), mpDevice->GetDescriptorHandleIncrementSize(D3D12_DESCRIPTOR_HEAP_TYPE_RTV)); }
 	inline ID3D12Resource*               GetCurrentBackBufferRenderTarget() const { return mRenderTargets[GetCurrentBackBufferIndex()]; }
